Clean searchable snapshots cache on startup
Today we empty the searchable snapshots cache when cleanly closing a
shard, but leak cache files in some cases involving an unclean shutdown.
Such leaks are not permanent, they are cleaned up on shard relocation or
deletion, but they still might last for arbitrarily long until that
happens. This commit introduces a cleanup process that runs during node
startup to catch such leaks sooner.

Also, today we permit searchable snapshots to be held on custom data
paths, and store the corresponding cache files within the custom
location. Supporting this feature would make the cleanup process
significantly more complicated since it would require each node to parse
the index metadata for the shards it held before shutdown. Yet, this
feature is undocumented and offers minimal benefits to searchable
snapshots. Therefore with this commit we forbid custom data paths for
searchable snapshot shards.

+++ x-pack/plugin/searchable-snapshots/src/main/java/org/elasticsearch/index/store/cache/CacheFile.java
@@ -55,6 +55,7 @@ public class CacheFile {
     private final SparseFileTracker tracker;
     private final String description;
     private final Path file;
+    private final boolean deleteOnEviction; // only false in tests
 
     private volatile Set<EvictionListener> listeners;
     private volatile boolean evicted;
@@ -62,7 +63,8 @@ public class CacheFile {
     @Nullable // if evicted, or there are no listeners
     private volatile FileChannel channel;
 
-    public CacheFile(String description, long length, Path file) {
+    public CacheFile(String description, long length, Path file, boolean deleteOnEviction) {
+        this.deleteOnEviction = deleteOnEviction;
         this.tracker = new SparseFileTracker(file.toString(), length);
         this.description = Objects.requireNonNull(description);
         this.file = Objects.requireNonNull(file);
@@ -168,10 +170,12 @@ public class CacheFile {
         assert evictionLock.isHeldByCurrentThread();
         assert listeners.isEmpty();
         assert channel == null;
-        try {
-            Files.deleteIfExists(file);
-        } catch (IOException e) {
-            throw new UncheckedIOException(e);
+        if (deleteOnEviction) {
+            try {
+                Files.deleteIfExists(file);
+            } catch (IOException e) {
+                throw new UncheckedIOException(e);
+            }
         }
     }
 
@@ -221,7 +225,7 @@ public class CacheFile {
             assert listeners != null;
             if (listeners.isEmpty()) {
                 assert channel == null;
-                assert evicted == false || refCounter.refCount() != 0 || Files.notExists(file);
+                assert evicted == false || refCounter.refCount() != 0 || Files.notExists(file) || deleteOnEviction == false;
             } else {
                 assert channel != null;
                 assert refCounter.refCount() > 0;
